EROPSPT-300: Use shared email package
