+++ src/main/kotlin/uk/gov/dluhc/registercheckerapi/config/SesEmailClientConfiguration.kt
@@ -4,7 +4,6 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import software.amazon.awssdk.regions.providers.DefaultAwsRegionProviderChain
 import software.amazon.awssdk.services.ses.SesClient
-import uk.gov.dluhc.emailnotifications.SesEmailClient
 
 /**
  * Configuration class exposing a configured email client bean to send emails via AWS's SES service.
@@ -16,16 +15,4 @@ class SesEmailClientConfiguration() {
         SesClient.builder()
             .region(DefaultAwsRegionProviderChain().region)
             .build()
-
-    @Bean
-    fun sesEmailClient(
-        sesClient: SesClient,
-        emailClientProperties: EmailClientProperties
-    ): SesEmailClient =
-        with(emailClientProperties) {
-            SesEmailClient(
-                sesClient = sesClient,
-                sender = sender,
-            )
-        }
 }
