+++ src/main/kotlin/uk/gov/dluhc/registercheckerapi/service/EmailService.kt
@@ -2,13 +2,13 @@ package uk.gov.dluhc.registercheckerapi.service
 
 import liquibase.repackaged.org.apache.commons.text.StringSubstitutor.replace
 import org.springframework.stereotype.Service
-import uk.gov.dluhc.emailnotifications.SesEmailClient
+import uk.gov.dluhc.email.EmailClient
 import uk.gov.dluhc.registercheckerapi.config.EmailContentConfiguration
 import uk.gov.dluhc.registercheckerapi.database.entity.RegisterCheckSummaryByGssCode
 
 @Service
 class EmailService(
-    private val emailClient: SesEmailClient,
+    private val emailClient: EmailClient,
     private val emailContentConfiguration: EmailContentConfiguration
 ) {
     fun sendRegisterCheckMonitoringEmail(
